% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcheck.r
\name{rcheck}
\alias{rcheck}
\title{Check a package}
\usage{
rcheck(path = getwd(), cleanup = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{cleanup}{[\code{logical(1L)}]\cr
Clean up the log directory, even if check failed or issued warnings. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Check a package located in \code{path} using \code{link[rcmdcheck]{rcmdcheck}}..
}
