% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclean.r
\name{rclean}
\alias{rclean}
\title{Remove temporary files from a package}
\usage{
rclean(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Cleanup temporary files for a package located in \code{path} using \code{link[devtools]{check}}.
}
