% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone.r
\name{rclone}
\alias{rclone}
\title{Clone a package}
\usage{
rclone(pkg, temp = FALSE)
}
\arguments{
\item{pkg}{[\code{character(1)}]\cr
String interpretable by \code{\link{stringToPackage}}.}

\item{temp}{[\code{logical(1)}]\cr
Clone to temp directory instead of current working directory?
Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Clones a package from GitHub.
Also works for CRAN packages (CRAN mirror is used in this case, \url{http://cran.github.io/}).
Great to have a quick peek with your favorite editor or to grep package sources.
}

