% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompile.r
\name{rcompile}
\alias{rcompile}
\title{Compile an R markdown or R noweb (Sweave) document}
\usage{
rcompile(files = character(0L), html = FALSE)
}
\arguments{
\item{files}{[\code{character}]\cr
List of knitr (.Rnw, .Rrst) and Rmarkdown (.Rmd) files to compile.}

\item{html}{[\code{logical(1)}]\cr
If \code{FALSE} (default), it generates a PDF.
If \code{TRUE}, the function will generate HTML output.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uses \code{rmarkdown} or \code{knitr} to compile its input argument
into an HTML or PDF file.
}
