% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.r
\name{rdoc}
\alias{rdoc}
\title{Create documentation for a package}
\usage{
rdoc(path = getwd(), staticdocs = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{staticdocs}{[\code{logical(1)}]\cr
Also run \code{staticdocs} (\url{https://github.com/hadley/staticdocs}) on the
generated man pages.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Documents a package located in \code{path} using \code{\link[roxygen2]{roxygen-package}}
or rather \code{link[devtools]{document}}.
}

