% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhub.r
\name{rhub}
\alias{rhub}
\title{Upload package to rhub}
\usage{
rhub(path = getwd(), platform)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{platform}{[\code{character(1)}]\cr
Check on the platform specified here.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uploads a package located in \code{path} to the rhub service.
}
