% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinstall.r
\name{rinstall}
\alias{rinstall}
\title{Installs R packages}
\usage{
rinstall(pkgs = ".", add = FALSE, ...)
}
\arguments{
\item{pkgs}{[\code{character}]\cr
Strings convertible by \code{\link{stringToPackage}}.
Default is the current directory: \dQuote{.}}

\item{add}{[\code{logical(1)}]\cr
Append packages to your config file in \code{\link{getConfigPath}}.?
Default is \code{FALSE}.}

\item{...}{[\code{any}]\cr
Passed to \code{remotes::install_*()}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Installs R packages from local directory, CRAN or Git.
}
