% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknit.r
\name{rknit}
\alias{rknit}
\title{Knit a document with knitr}
\usage{
rknit(file, html = FALSE)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Document to knit.}

\item{html}{[\code{logical(1)}]\cr
\code{FALSE} (default), it generates a PDF.
\code{TRUE}, the function will generate HTML output.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Calls \code{\link[knitr]{knit2pdf}} on a file.
}
