% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmake.r
\name{rmake}
\alias{rmake}
\title{Make and install a package}
\usage{
rmake(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Updates the documentation and then installs the package located at \code{path}, using
\code{link[roxygen2]{roxygenize}}, \code{link[remotes]{install_deps}} and \code{link[remotes]{install_local}}.
}
