% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.r
\name{rmarkdown}
\alias{rmarkdown}
\title{Render a markdown document with rmarkdown}
\usage{
rmarkdown(file)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Document to render.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Calls \code{\link[rmarkdown]{render}} on a file.
}
