% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgdown.r
\name{rpkgdown}
\alias{rpkgdown}
\title{Create static documentation with pkgdown}
\usage{
rpkgdown(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Build static documentation with \url{https://github.com/hadley/pkgdown} in the
\dQuote{doc} subdirectory of the package root.
}
