% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrelease.r
\name{rrelease}
\alias{rrelease}
\title{Release a package}
\usage{
rrelease(path = getwd(), directsubmit = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{directsubmit}{[\code{logical(1)}]\cr
Submit package without questions and pause to CRAN.
Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Release the package located at \code{path}, using \code{link[devtools]{release}}.
}
