% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rremove.R
\name{rremove}
\alias{rremove}
\title{Remove R packages}
\usage{
rremove(pkgs = character(0L))
}
\arguments{
\item{pkgs}{[\code{character}]\cr
Strings convertible by \code{\link{stringToPackage}}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Removes (uninstalls) R packages.
}
