% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshine.r
\name{rshine}
\alias{rshine}
\title{Run a shiny app}
\usage{
rshine(path = ".", port = NULL)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{port}{[\code{integer(1)}]\cr
Port for \code{\link[shiny]{runApp}}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Runs a shiny app.
}
