% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspell.r
\name{rspell}
\alias{rspell}
\title{Check spelling in generated Rd files}
\usage{
rspell(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Check spelling in generated Rd files
}

