% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.r
\name{rt}
\alias{rt}
\title{Maintainace operations for rt}
\usage{
rt(init = FALSE, edit = FALSE)
}
\arguments{
\item{init}{[\code{logical(1)}]\cr
Initialize for first use: create directory \code{\link{getConfigPath}} and
create an empty package collection file \code{getConfigPath("packages")}.}

\item{edit}{[\code{logical(1)}]\cr
Edit the \dQuote{packages} file?}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Maintainace operations for rt
}
