% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtest.r
\name{rtest}
\alias{rtest}
\title{Run unit tests for a package}
\usage{
rtest(path = getwd(), filter = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{filter}{[\code{FALSE} || \code{character(1)}]\cr
Filter tests (on a file name basis) using this pattern.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Tests a package located in \code{path} using \code{\link[devtools]{test}}.
}

