% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtstchk.r
\name{rtstchk}
\alias{rtstchk}
\title{Run unit tests and check a package}
\usage{
rtstchk(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
First test than check a package located in \code{path}
 using \code{\link[devtools]{test}} and \code{link[devtools]{check}}.
}

