% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupdate.r
\name{rupdate}
\alias{rupdate}
\title{Update packages}
\usage{
rupdate(rebuilt = FALSE)
}
\arguments{
\item{rebuilt}{[\code{logical(1)}]\cr
Rebuilt R packages which are built using a different version of R.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Update packages. This includes multiple steps:
\itemize{
  \item{1.}{All outdated CRAN packages will be updated.}
  \item{2.}{If \code{rebuilt} is set to \code{TRUE}, all packages which
     are built under a different R version will be rebuilt with the current.}
  \item{3.}{CRAN packages which are listed in \code{~/.rt/packages} but are not
    found on the system will be installed.}
  \item{4.}{All packages with a git source will be updated and installed if necessary.}
}
}

