% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvignette.r
\name{rvignettes}
\alias{rvignettes}
\title{Build vignettes}
\usage{
rvignettes(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Build vignettes of the package located in \code{path} using \code{\link[=devtools]{build_vignette}}.
}

