% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwinbuild.r
\name{rwinbuild}
\alias{rwinbuild}
\title{Upload package to winbuilder}
\usage{
rwinbuild(path = getwd(), devel = FALSE, oldrel = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{devel}{[\code{logical(1)}]\cr
Upload to check against the R-devel branch instead of the stable branch.}

\item{oldrel}{[\code{logical(1)}]\cr
Upload to check against the R-oldrelease branch instead of the stable branch.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uploads a package located in \code{path} to the winbuilder service.
Uses \code{\link[devtools]{check_win_release}}, \code{\link[devtools]{check_win_devel}} or \code{\link[devtools]{check_win_oldrelease}} internally.
}
