% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringToPackage.r
\name{stringToPackage}
\alias{stringToPackage}
\title{Convert a string to an object of type 'Package'}
\usage{
stringToPackage(pkg)
}
\arguments{
\item{pkg}{[\code{character(1)} | \code{NULL}]\cr
String to convert to an object of class Package.}
}
\description{
This function converts the string to one of the following objects:
\itemize{
  \item{LocalPackage: }{\code{pkg} points to an existing directory.}
  \item{GitPackage: }{\code{pkg} starts with \dQuote{https://} or \dQuote{git} and ends with \dQuote{.git}.}
  \item{GitHubPackage: }{\code{pkg} matches the pattern \dQuote{[user]/[repo]}.}
  \item{CranPackage: }{\code{pkg} is a single string without any special characters except dots ([a-zA-Z0-9.]).}
}
}

