% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.r
\name{getConfigPath}
\alias{getConfigPath}
\title{Path to rt configuration}
\usage{
getConfigPath(element = ".")
}
\arguments{
\item{element}{[\code{character(1)}]\cr
Which file whithin the config dir to acess.
Use \dQuote{.} for the directory itself.
Uses \code{\link[rappdirs]{user_config_dir}} to determine the path.}
}
\value{
Returns the path as character.
}
\description{
Returns the path to the rt configuration files
}
