% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbuild.r
\name{rbuild}
\alias{rbuild}
\title{Built a package}
\usage{
rbuild(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Built a package located in \code{path} using \code{link[devtools]{build}}.
}

