% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompile.r
\name{rcompile}
\alias{rcompile}
\title{Compile an R markdown or R noweb (Sweave) document}
\usage{
rcompile(file = character(0L), html = FALSE)
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uses \code{rmarkdown} or \code{knitr} to compile its input argument
into an HTML or PDF file.
}

