% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcov.r
\name{rcov}
\alias{rcov}
\title{Test the coverage of a package}
\usage{
rcov(path = getwd(), shine = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{shine}{[\code{logical(1)}]\cr
Use the function \code{link[covr]{shine}} to display the coverage information.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uses the \code{covr} package (\url{https://github.com/jimhester/covr}) to
test the code coverage.
}

