% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.r
\name{rdoc}
\alias{rdoc}
\title{Create documentation for a package}
\usage{
rdoc(path = getwd())
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Documents a package located in \code{path} using \code{\link[roxygen2]{roxygen2-package}}
or rather \code{link[devtools]{document}}.
}
