% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinstall.r
\name{rinstall}
\alias{rinstall}
\title{Installs R packages}
\usage{
rinstall(pkgs = character(0L), add = FALSE, noquick = FALSE)
}
\arguments{
\item{pkgs}{[\code{character}]\cr
Strings convertible by \code{\link{stringToPackage}}.}

\item{add}{[\code{logical(1)}]\cr
Append packages to your \code{~/.rt/packages} file?
Default is \code{FALSE}.}

\item{noquick}{[\code{logical(1)}]\cr
Switch off devtools quick installation for git packages.
See \code{?devtools::install}.
Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Installs R packages from local directory, CRAN or Git.
}

