% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknit.r
\name{rknit}
\alias{rknit}
\title{Knit a document with knitr}
\usage{
rknit(file)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Document to knit.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Calls \code{\link[knitr]{knit2pdf}} on a file.
}
