% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknit.r
\name{rknit}
\alias{rknit}
\title{Knit a document with knitr}
\usage{
rknit(file, output_format = NULL)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Document to knit.}

\item{output_format}{[\code{character(1)}]\cr
\code{pdf_document}, \code{html_document} or \code{word_document}.
For Rnw-files output will be always PDF which is generated by running \code{\link[knitr]{knit2pdf}}.
For Rmd-files the YAML header information for \code{output} will be taken into account as default but can be overridden by this parameter.
Default is HTML.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Calls \code{\link[knitr]{knit2pdf}} (Rnw) or \code{\link[rmarkdown]{render}} (Rmd) on a file.
}
