% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmake.r
\name{rmake}
\alias{rmake}
\title{Make and install a package}
\usage{
rmake(path = getwd(), deps = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{deps}{[\code{logical(1)}]\cr
Also install all dependencies, including suggests? Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Updates the documenation and then installs the package located at \code{path}, using
\code{link[devtools]{document}}, \code{link[devtools]{install_deps}} and \code{link[devtools]{install}}.
}

