% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.r
\name{rmarkdown}
\alias{rmarkdown}
\title{Render a markdown document with rmarkdown}
\usage{
rmarkdown(file, html = FALSE)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Document to render.}

\item{html}{[\code{logical(1)}]\cr
\code{FALSE} (default), it generates a PDF.
\code{TRUE}, the function will generate HTML output.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Calls \code{\link[rmarkdown]{render}} on a file.
}
