% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rt.r
\name{rt}
\alias{rt}
\title{Maintainace operations for rt}
\usage{
rt(init = FALSE, cleanup = FALSE)
}
\arguments{
\item{init}{[\code{logical(1)}]\cr
Initialize for first use: create directory \code{~/.rt} and
create an empty package collection file \code{~/.rt/packages}.}

\item{cleanup}{[\code{logical(1)}]\cr
Remove git repositories in \code{~/.rt/git} which are not
referenced in the collection file.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Maintainace operations for rt
}

