% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtest.r
\name{rtest}
\alias{rtest}
\title{Run unit tests for a package}
\usage{
rtest(path = getwd(), filter = NULL)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{filter}{[\code{character(1)}]\cr
Filter tests (on a file name basis) using this pattern.
Default is \dQuote{NULL}, which means no filter is applied.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Tests a package located in \code{path} using tests in \code{tests/testthat} via
 \code{\link[testthat]{test_dir}}.
}
