% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupdate.r
\name{rupdate}
\alias{rupdate}
\title{Update packages}
\usage{
rupdate(rebuild = FALSE, neverupgrade = FALSE)
}
\arguments{
\item{rebuild}{[\code{logical(1)}]\cr
Rebuild R packages which are built using a different version of R.}

\item{neverupgrade}{[code{logical(1)}]\cr
Passed to \code{\link[remotes]{update_packages}} as \code{upgrade = ifelse(neverupgrade, "never", "always")}.
Default is \dQuote{FALSE} so all dependencies are upgraded.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Update packages. This includes multiple steps:
\itemize{
  \item{1.}{All outdated CRAN packages will be updated.}
  \item{2.}{If \code{rebuild} is set to \code{TRUE}, all packages which
     are built under a different R version will be rebuilt with the current.}
  \item{3.}{CRAN packages which are listed in \code{~/.rt/packages} but are not
    found on the system will be installed.}
  \item{4.}{All packages with a git source listed in \code{~/.rt/packages} will be updated and
    re-installed if necessary.}
}
The collection file \code{~/.rt/packages} expects one package per line specified in a format
which is parseable by \code{\link{stringToPackage}}. Leading and trailing whitespace characters
will automatically be trimmed. Empty lines and lines starting with a \dQuote{#} will be ignored.
}
