% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwinbuild.r
\name{rwinbuild}
\alias{rwinbuild}
\title{Upload package to winbuilder}
\usage{
rwinbuild(path = getwd(), devel = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
If no path to a DESCRIPTION is given, the package looks for a DESCRIPTION in
the current directory and up to two parent directories.}

\item{devel}{[\code{logical(1)}]\cr
Upload to check against the R-devel branch instead of the stable branch.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Uploads a package located in \code{path} to the winbuilder service. Uses
\code{\link[devtools]{build_win}} internally.
}

