% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringToPackage.r
\name{stringToPackage}
\alias{stringToPackage}
\title{Convert a string to an object of type 'Package'}
\usage{
stringToPackage(pkg)
}
\arguments{
\item{pkg}{[\code{character(1)} | \code{NULL}]\cr
String to convert to an object of class Package.}
}
\description{

}
